Playing_With_Characters.c
Playing With Characters :

Task

You have to print the character, , in the first line. Then print in next line. In the last line print the sentence, .

Input Format

First, take a character, as input.
Then take the string, as input.
Lastly, take the sentence as input.

Output Format

Print three lines of output. The first line prints the character, .
The second line prints the string, .
The third line prints the sentence, .